Map Parsing

+++ includes/cube3d_vector.h
@@ -6,7 +6,7 @@
 /*   By: gozsertt <gozsertt@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/31 14:15:36 by gozsertt          #+#    #+#             */
-/*   Updated: 2019/12/31 14:23:59 by gozsertt         ###   ########.fr       */
+/*   Updated: 2020/01/10 15:43:41 by gozsertt         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,12 +15,12 @@
 
 typedef struct s_vector
 {
-	int	x;
-	int	y;
+	float	x;
+	float	y;
 }			t_vector;
 
-t_vector	*malloc_vector(int x, int y);
-t_vector	create_vector(int x, int y);
+t_vector	*malloc_vector(float x, float y);
+t_vector	create_vector(float x, float y);
 void		destroy_vector(t_vector to_destroy);
 void		free_vector(t_vector *to_free);
