+++ srcs/structure/vector.c
@@ -6,13 +6,13 @@
 /*   By: gozsertt <gozsertt@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/31 14:20:54 by gozsertt          #+#    #+#             */
-/*   Updated: 2020/01/02 12:20:54 by gozsertt         ###   ########.fr       */
+/*   Updated: 2020/01/10 15:43:41 by gozsertt         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cube3d.h"
 
-t_vector	*malloc_vector(int x, int y)
+t_vector	*malloc_vector(float x, float y)
 {
 	t_vector *vector;
 
@@ -23,7 +23,7 @@ t_vector	*malloc_vector(int x, int y)
 	return (vector);
 }
 
-t_vector	create_vector(int x, int y)
+t_vector	create_vector(float x, float y)
 {
 	t_vector vector;
